+++ src/js/api-service.js
@@ -40,6 +40,14 @@ class ApiService {
         return data.results;
       });
   }
+  getMovieData(id) {
+    return fetch(`${this.BASE_URL}/movie/${id}?api_key=${this.API_KEY}`).then(data => {
+      if (data.ok) {
+        return data.json();
+      }
+      return Promise.reject(new Error('Error'));
+    });
+  }
 
   // Search & fetch Movies with header input
   // searchMovies() {
